Added features and fixed a bug.
When a user correctly guess a municipality a new image of Denmark with the municipality superimposed will be shown.

The user can hit enter when typing instead of having to press the guess button.

The list of municipalities has been reset to start on 2022-05-06.

The list of municipalities in the auto complete is now sorted alphabetically.